Updated .gitignore and added --run feature.
Added .gitignore to avoid accidentally uploading .csv files, which could contain personal information and should not be part of the code repository.

In addition, the run feature is designed to allow a user with a CSV with only 'primaryEmail' to immediately suspend all accounts included in the CSV.

I also added an additional GAM argument for suspended accounts. `gal off` removes the user account from the Global Address List, which keeps a tidier Directory structure, which can be found at https://contacts.google.com.

Changes to be committed:
	modified:   .gitignore
	modified:   csv_suspend.py